+++ autofit/mapper/__init__.py
@@ -1,5 +1,5 @@
 from autofit.mapper.prior_model.prior_model import PriorModel, CollectionPriorModel, \
     AbstractPriorModel, AnnotationPriorModel
 from .model_mapper import ModelMapper
-from .prior import Prior, UniformPrior, GaussianPrior, Constant, DeferredArgument, \
-    LogUniformPrior, AttributeNameValue
+from .prior import Prior, UniformPrior, GaussianPrior, Constant, LogUniformPrior, \
+    AttributeNameValue
