+++ test/mapper/test_deference.py
@@ -1,5 +1,6 @@
 import pytest
 
+import autofit.mapper.prior_model.deferred
 from autofit import exc
 from autofit.mapper import model_mapper as mm
 from autofit.mapper import prior as p
@@ -11,7 +12,7 @@ from test.mock import DeferredClass
 @pytest.fixture(name="deferred_instance")
 def make_deferred_instance():
     prior_model = pm.PriorModel(mock.SimpleClass)
-    prior_model.two = p.DeferredArgument()
+    prior_model.two = autofit.mapper.prior_model.deferred.DeferredArgument()
 
     return prior_model.instance_for_arguments(
         {
@@ -22,7 +23,8 @@ def make_deferred_instance():
 
 class TestCase:
     def test_is_deferred(self, deferred_instance):
-        assert isinstance(deferred_instance, pm.DeferredInstance)
+        assert isinstance(deferred_instance,
+                          autofit.mapper.prior_model.deferred.DeferredInstance)
 
     def test_instantiate(self, deferred_instance):
         instance = deferred_instance(two=2.0)
@@ -42,7 +44,8 @@ class TestCase:
 
         deferred_instance = mapper.instance_from_unit_vector([1.0]).DeferredClass
 
-        assert isinstance(deferred_instance, pm.DeferredInstance)
+        assert isinstance(deferred_instance,
+                          autofit.mapper.prior_model.deferred.DeferredInstance)
 
         instance = deferred_instance(two=2.0)
